docker/backend-requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
psycopg2-binary==2.9.9
bcrypt==4.1.2
boto3==1.34.0
pillow==10.4.0
requests==2.31.0
pypdf==4.0.1
python-multipart==0.0.9
